+++ resources/views/sites.blade.php
@@ -154,8 +154,8 @@
     //Get DT Data
     getData('/api/sites');
 
-    //Datatable
-    function dtRender() {
+    // Render Make
+    function renderMake() {
         $('#dt').DataTable( {
             'processing': true,
             'data': JSON.parse(localStorage.getItem('dtdata')),
@@ -221,18 +221,11 @@
                     $('#loadingdelete').removeClass('d-none');
                 },
                 complete: function(data) {
-                    setTimeout(function() {
-                        $('#dt').DataTable().clear().destroy();
-                    }, 4500);
-                    setTimeout(function() {
-                        getData('/api/sites',false);
-                    }, 6000);
-                    setTimeout(function() {
-                        $('#deleteSiteModal').modal('toggle');
-                        $('#deletesitedomain').html('');
-                        $('#deletesiteid').val('');
-                        $('#loadingdelete').addClass('d-none');
-                    }, 6500);
+                    getData('/api/sites',false);
+                    $('#deleteSiteModal').modal('toggle');
+                    $('#deletesitedomain').html('');
+                    $('#deletesiteid').val('');
+                    $('#loadingdelete').addClass('d-none');
                 },
             });
         });